Remove statsProcessors
