+++ src/types/domain/Champion.ts
@@ -1,7 +1,7 @@
 export type Champion = {
-    name: string;
-    wins: number;
-    losses: number;
-    winPercentage: number;
-    totalGames: number;
-}
+  name: string;
+  wins: number;
+  losses: number;
+  winPercentage: number;
+  totalGames: number;
+};
